fix over sample util tests and rename
