+++ test_autoarray/operators/over_sample/test_over_sample_util.py
@@ -367,3 +367,81 @@ def test__grid_2d_slim_over_sampled_via_mask_from():
     assert grid[0:4] == pytest.approx(
         np.array([[1.75, -0.5], [1.75, 2.5], [0.25, -0.5], [0.25, 2.5]]), 1e-4
     )
+
+
+def test__from_manual_adapt_radial_bin():
+    mask = aa.Mask2D.circular(shape_native=(5, 5), pixel_scales=2.0, radius=3.0)
+
+    grid = aa.Grid2D.from_mask(mask=mask)
+
+    sub_size = aa.util.over_sample.over_sample_size_via_radial_bins_from(
+        grid=grid, sub_size_list=[8, 4, 2], radial_list=[1.5, 2.5]
+    )
+    assert sub_size.native == pytest.approx(
+        np.array(
+            [
+                [0, 0, 0, 0, 0],
+                [0, 2, 4, 2, 0],
+                [0, 4, 8, 4, 0],
+                [0, 2, 4, 2, 0],
+                [0, 0, 0, 0, 0],
+            ]
+        ),
+        1.0e-4,
+    )
+
+
+def test__from_manual_adapt_radial_bin__centre_list_input():
+    mask = aa.Mask2D.circular(shape_native=(5, 5), pixel_scales=2.0, radius=3.0)
+
+    grid = aa.Grid2D.from_mask(mask=mask)
+
+    sub_size = aa.util.over_sample.over_sample_size_via_radial_bins_from(
+        grid=grid,
+        sub_size_list=[8, 4, 2],
+        radial_list=[1.5, 2.5],
+        centre_list=[(0.0, -2.0), (0.0, 2.0)],
+    )
+
+    assert sub_size.native == pytest.approx(
+        np.array(
+            [
+                [0, 0, 0, 0, 0],
+                [0, 4, 2, 4, 0],
+                [0, 8, 4, 8, 0],
+                [0, 4, 2, 4, 0],
+                [0, 0, 0, 0, 0],
+            ]
+        ),
+        1.0e-4,
+    )
+
+
+def test__from_adapt():
+    mask = aa.Mask2D(
+        mask=[[True, True, True], [True, False, False], [True, True, False]],
+        pixel_scales=1.0,
+    )
+
+    data = aa.Array2D(values=[1.0, 2.0, 3.0], mask=mask)
+    noise_map = aa.Array2D(values=[1.0, 2.0, 1.0], mask=mask)
+
+    sub_size = aa.util.over_sample.over_sample_size_via_adapt_from(
+        data=data,
+        noise_map=noise_map,
+        signal_to_noise_cut=1.5,
+        sub_size_lower=2,
+        sub_size_upper=4,
+    )
+
+    assert sub_size == pytest.approx([2, 2, 4], 1.0e-4)
+
+    sub_size = aa.util.over_sample.over_sample_size_via_adapt_from(
+        data=data,
+        noise_map=noise_map,
+        signal_to_noise_cut=0.5,
+        sub_size_lower=2,
+        sub_size_upper=4,
+    )
+
+    assert sub_size == pytest.approx([4, 4, 4], 1.0e-4)
