+++ test_autoarray/operators/over_sample/test_over_sampler.py
@@ -51,84 +51,6 @@ def test__sub_pixel_areas():
     assert areas == pytest.approx([1.0, 0.25, 0.25, 0.25, 0.25], 1.0e-4)
 
 
-def test__from_manual_adapt_radial_bin():
-    mask = aa.Mask2D.circular(shape_native=(5, 5), pixel_scales=2.0, radius=3.0)
-
-    grid = aa.Grid2D.from_mask(mask=mask)
-
-    over_sampling = aa.OverSampling.from_radial_bins(
-        grid=grid, sub_size_list=[8, 4, 2], radial_list=[1.5, 2.5]
-    )
-    assert over_sampling.sub_size.native == pytest.approx(
-        np.array(
-            [
-                [0, 0, 0, 0, 0],
-                [0, 2, 4, 2, 0],
-                [0, 4, 8, 4, 0],
-                [0, 2, 4, 2, 0],
-                [0, 0, 0, 0, 0],
-            ]
-        ),
-        1.0e-4,
-    )
-
-
-def test__from_manual_adapt_radial_bin__centre_list_input():
-    mask = aa.Mask2D.circular(shape_native=(5, 5), pixel_scales=2.0, radius=3.0)
-
-    grid = aa.Grid2D.from_mask(mask=mask)
-
-    over_sampling = aa.OverSampling.from_radial_bins(
-        grid=grid,
-        sub_size_list=[8, 4, 2],
-        radial_list=[1.5, 2.5],
-        centre_list=[(0.0, -2.0), (0.0, 2.0)],
-    )
-
-    assert over_sampling.sub_size.native == pytest.approx(
-        np.array(
-            [
-                [0, 0, 0, 0, 0],
-                [0, 4, 2, 4, 0],
-                [0, 8, 4, 8, 0],
-                [0, 4, 2, 4, 0],
-                [0, 0, 0, 0, 0],
-            ]
-        ),
-        1.0e-4,
-    )
-
-
-def test__from_adapt():
-    mask = aa.Mask2D(
-        mask=[[True, True, True], [True, False, False], [True, True, False]],
-        pixel_scales=1.0,
-    )
-
-    data = aa.Array2D(values=[1.0, 2.0, 3.0], mask=mask)
-    noise_map = aa.Array2D(values=[1.0, 2.0, 1.0], mask=mask)
-
-    over_sampling = aa.OverSampling.from_adapt(
-        data=data,
-        noise_map=noise_map,
-        signal_to_noise_cut=1.5,
-        sub_size_lower=2,
-        sub_size_upper=4,
-    )
-
-    assert over_sampling.sub_size == pytest.approx([2, 2, 4], 1.0e-4)
-
-    over_sampling = aa.OverSampling.from_adapt(
-        data=data,
-        noise_map=noise_map,
-        signal_to_noise_cut=0.5,
-        sub_size_lower=2,
-        sub_size_upper=4,
-    )
-
-    assert over_sampling.sub_size == pytest.approx([4, 4, 4], 1.0e-4)
-
-
 def test__sub_fraction():
     mask = aa.Mask2D(
         mask=[[False, False], [True, True]],
